Modification to support binary relocation when the FILESDIR macro is a relative path.
A second macro BINDIR can be used in combination to indicate an optional binary folder
where the program is installed (e.g. the bin part from /usr/bin).

When FILESDIR is a relative path, the computed files dir will be the concatenation of
the executable parent path [- BINDIR>] + <FILESDIR>. If FILESDIR is an absolute path,
it is used unmodified.

This is a possible solution to https://github.com/danmar/cppcheck/pull/4424.